+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Static profile components.
 - Example project profile.
+- Selected state for tab bar.
 
 ## 0.0.1 - 2017-09-21
 ### Added